Use ParticipantsSearchResults in SetContacts
