+++ src/components/RightSidebar/Participants/ParticipantsList/ParticipantsList.vue
@@ -21,60 +21,24 @@
 
 <template>
 	<div>
-		<ul v-if="(!loading || addOnClick) && !noResults">
+		<ul>
 			<Participant
-				v-for="participant in participants"
-				:key="participant.userId + participant.sessionId"
-				:participant="participant"
+				v-for="item in items"
+				:key="item.id"
+				:participant="item"
 				@clickParticipant="handleClickParticipant" />
-			<!-- 'search for more' empty content to display at the end of the
-			participants list, this is useful in case the participants list is used
-			to display the results of a search. Upon clicking on it, an event is
-			emitted to the parent component in order to be able to focus on it's
-			input field -->
-			<li
-				v-if="displaySearchHint"
-				class="participants-list__hint"
-				@click="handleClickHint">
-				<div class="icon-contacts-dark set-contacts__icon" />
-				<p>
-					{{ t('spreed', 'Search for more contacts') }}
-				</p>
-			</li>
 		</ul>
-		<template v-if="loading">
-			<template v-if="addOnClick">
-				<LoadingParticipant
-					v-for="n in dummyParticipants"
-					:key="n" />
-			</template>
-			<template v-else>
-				<div class="icon-loading participants-list__icon" />
-				<p class="participants-list__warning">
-					{{ t('spreed', 'Contacts loading') }}
-				</p>
-			</template>
-		</template>
-		<template v-if="noResults">
-			<div class="icon-category-search participants-list__icon" />
-			<p class="participants-list__warning">
-				{{ t('spreed', 'No results') }}
-			</p>
-		</template>
 	</div>
 </template>
 
 <script>
 
-import LoadingParticipant from './Participant/LoadingParticipant'
 import Participant from './Participant/Participant'
-import Vue from 'vue'
 
 export default {
 	name: 'ParticipantsList',
 
 	components: {
-		LoadingParticipant,
 		Participant,
 	},
 
@@ -86,43 +50,6 @@ export default {
 			type: Array,
 			required: true,
 		},
-		/**
-		 * If true, clicking the participant will add it to to the current conversation.
-		 * This behavior is used in the right sidebar for already existing conversations.
-		 * If false, clicking on the participant will add the participant to the
-		 * `selectedParticipants` array in the data.
-		 */
-		addOnClick: {
-			type: Boolean,
-			default: true,
-		},
-		/**
-		 * Display loading state instead of list.
-		 */
-		loading: {
-			type: Boolean,
-			default: false,
-		},
-		/**
-		 * Display no-results state instead of list.
-		 */
-		noResults: {
-			type: Boolean,
-			default: false,
-		},
-		/**
-		 * Display 'search for more' empty content at the end of the list.
-		 */
-		displaySearchHint: {
-			type: Boolean,
-			default: false,
-		},
-	},
-
-	data() {
-		return {
-			selectedParticipants: [],
-		}
 	},
 
 	computed: {
@@ -134,64 +61,12 @@ export default {
 			const dummies = 6 - this.items.length
 			return dummies > 0 ? dummies : 0
 		},
-
-		/**
-		 * Creates a new array that combines the items (participants received as a prop)
-		 * with the current selectedParticipants so that each participant in the returned
-		 * array has a new 'selected' boolean key.
-		 * @returns {array} An array of 'participant' objects
-		 */
-		participants() {
-			/**
-			 * Compute this only in the new group conversation form.
-			 */
-			if (this.addOnClick === false) {
-				if (this.items !== []) {
-					const participants = this.items.slice()
-					participants.forEach(item => {
-						if (this.selectedParticipants.indexOf(item) !== -1) {
-							Vue.set(item, 'selected', true)
-						} else {
-							Vue.set(item, 'selected', false)
-						}
-					})
-					return participants
-				} else {
-					return []
-				}
-			} else {
-				return this.items
-			}
-		},
 	},
 
 	methods: {
 		async handleClickParticipant(participant) {
-			if (this.addOnClick) {
-				this.$emit('click', participant)
-			} else {
-				/**
-				 * Remove the clicked participant from the selected participants list
-				 */
-				if (this.selectedParticipants.indexOf(participant) !== -1) {
-					this.selectedParticipants = this.selectedParticipants.filter((selectedParticipant) => {
-						return selectedParticipant.id !== participant.id
-					})
-					this.$emit('updateSelectedParticipants', this.selectedParticipants)
-				} else {
-					/**
-					 * Add the clicked participant from the selected participants list
-					 */
-					this.selectedParticipants = [...this.selectedParticipants, participant]
-					this.$emit('updateSelectedParticipants', this.selectedParticipants)
-				}
-			}
-
+			this.$emit('click', participant)
 		},
-		handleClickHint() {
-			this.$emit('clickSearchHint')
-		},
-
 	},
 }
 </script>
